Added HashingJDAO and RollingHashingJDAO
